requirements/requirements.txt
appnope==0.1.0
backcall==0.1.0
backoff==1.10.0
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.1
Django==2.2.7
django-redis==4.8.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
elastic-apm==4.2.2
idna==2.8
ipdb==0.12.2
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.4.2
python-dotenv==0.10.3
pytz==2019.3
raven==6.10.0
redis==3.3.11
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7